Fixes for refset in 19 version bundle.

+++ rule-engine-integration/src/main/java/org/ihtsdo/rules/test/DomainModelTestCheck.java
@@ -131,10 +131,10 @@ public class DomainModelTestCheck extends TestCase {
 			config = tf.newAceFrameConfig();
 			config.addViewPosition(tf.newPosition(
 					tf.getPath(new UUID[] {UUID.fromString("2faa9260-8fb2-11db-b606-0800200c9a66")}), 
-					Integer.MAX_VALUE));
+					Long.MAX_VALUE));
 			config.addViewPosition(tf.newPosition(
 					tf.getPath(new UUID[] {UUID.fromString("8c230474-9f11-30ce-9cad-185a96fd03a2")}), 
-					Integer.MAX_VALUE));
+					Long.MAX_VALUE));
 			config.addEditingPath(tf.getPath(new UUID[] {UUID.fromString("8c230474-9f11-30ce-9cad-185a96fd03a2")}));
 			config.getDescTypes().add(ArchitectonicAuxiliary.Concept.FULLY_SPECIFIED_DESCRIPTION_TYPE.localize().getNid());
 			config.getDescTypes().add(ArchitectonicAuxiliary.Concept.PREFERRED_DESCRIPTION_TYPE.localize().getNid());
