+++ wb-api/src/main/java/org/dwfa/ace/task/CloneConcept.java
@@ -83,7 +83,7 @@ public class CloneConcept extends AbstractTask {
             I_ConfigAceFrame config = (I_ConfigAceFrame) worker.readAttachement(WorkerAttachmentKeys.ACE_FRAME_CONFIG.name());
             Set<PositionBI> positionSet = new HashSet<PositionBI>();
             for (PathBI path : config.getEditingPathSet()) {
-            	positionSet.add(tf.newPosition(path, Integer.MAX_VALUE));
+            	positionSet.add(tf.newPosition(path, Long.MAX_VALUE));
             }
             PositionSetReadOnly clonePositions = new PositionSetReadOnly(positionSet);
             I_HostConceptPlugins host = (I_HostConceptPlugins) worker.readAttachement(WorkerAttachmentKeys.I_HOST_CONCEPT_PLUGINS.name());
