+++ wb-api/src/main/java/org/dwfa/ace/task/classify/SnoTaskCheckRedundant.java
@@ -390,7 +390,7 @@ public class SnoTaskCheckRedundant extends AbstractTask implements ActionListene
             cEditPathNid = cEditPathObj.getConceptNid();
             cEditIPath = tf.getPath(cEditPathObj.getUids());
             cEditPathPos = new ArrayList<PositionBI>();
-            cEditPathPos.add(tf.newPosition(cEditIPath, Integer.MAX_VALUE));
+            cEditPathPos.add(tf.newPosition(cEditIPath, Long.MAX_VALUE));
             addPathOrigins(cEditPathPos, cEditIPath);
 
             // GET ALL CLASSIFER_PATH ORIGINS
@@ -403,7 +403,7 @@ public class SnoTaskCheckRedundant extends AbstractTask implements ActionListene
             cClassPathNid = cClassPathObj.getConceptNid();
             cClassIPath = tf.getPath(cClassPathObj.getUids());
             cClassPathPos = new ArrayList<PositionBI>();
-            cClassPathPos.add(tf.newPosition(cClassIPath, Integer.MAX_VALUE));
+            cClassPathPos.add(tf.newPosition(cClassIPath, Long.MAX_VALUE));
             addPathOrigins(cClassPathPos, cClassIPath);
 
         } catch (TerminologyException e) {
