+++ wb-api/src/main/java/org/dwfa/ace/task/classify/SnorocketTaskOld.java
@@ -1075,7 +1075,7 @@ public class SnorocketTaskOld extends AbstractTask implements ActionListener {
             cEditPathNid = cEditPathObj.getConceptNid();
             cEditIPath = tf.getPath(cEditPathObj.getUids());
             cEditPathPos = new ArrayList<PositionBI>();
-            cEditPathPos.add(tf.newPosition(cEditIPath, Integer.MAX_VALUE));
+            cEditPathPos.add(tf.newPosition(cEditIPath, Long.MAX_VALUE));
             getPathOrigins(cEditPathPos, cEditIPath);
 
             // GET ALL CLASSIFER_PATH ORIGINS
@@ -1088,7 +1088,7 @@ public class SnorocketTaskOld extends AbstractTask implements ActionListener {
             cClassPathNid = cClassPathObj.getConceptNid();
             cClassIPath = tf.getPath(cClassPathObj.getUids());
             cClassPathPos = new ArrayList<PositionBI>();
-            cClassPathPos.add(tf.newPosition(cClassIPath, Integer.MAX_VALUE));
+            cClassPathPos.add(tf.newPosition(cClassIPath, Long.MAX_VALUE));
             getPathOrigins(cClassPathPos, cClassIPath);
 
         } catch (TerminologyException e) {
