+++ wb-bdb-mojo/src/main/java/org/ihtsdo/bdb/mojo/DefaultConfig.java
@@ -240,7 +240,7 @@ public class DefaultConfig {
             PathBI editPath = tf.getPath(Concept.ARCHITECTONIC_BRANCH.getUids());
             // activeConfig.addEditingPath(editPath);
 
-            PositionBI viewPosition = tf.newPosition(editPath, Integer.MAX_VALUE);
+            PositionBI viewPosition = tf.newPosition(editPath, Long.MAX_VALUE);
             Set<PositionBI> viewSet = new HashSet<PositionBI>();
             viewSet.add(viewPosition);
             activeConfig.setViewPositions(viewSet);
