+++ wb-gui/src/main/java/org/dwfa/ace/gui/concept/ConflictPanel.java
@@ -432,7 +432,7 @@ public class ConflictPanel extends JPanel implements ActionListener {
             AceLog.getEditLog().fine("processing editPath: " + editPath);
         }
         Set<I_Position> positions = new HashSet<I_Position>();
-        positions.add(new Position(Integer.MAX_VALUE, editPath));
+        positions.add(new Position(Long.MAX_VALUE, editPath));
         List<I_ConceptAttributeVersioned> attributeList = new ArrayList<I_ConceptAttributeVersioned>();
         attributeList.add(cb.getConceptAttributes());
